Move Benchmark_buildEntitiesRequest_geohashArrays to feature_retriever_bench_test.go
